refactor structure again, final
