+++ .gitignore
@@ -2,7 +2,8 @@
 phoenix-logs/
 .idea/
 test.py
-models/
+ml/data/models/*
+!ml/data/models/b0_new
 
 # Byte-compiled / optimized / DLL files
 __pycache__/
