+++ game/src/core/shanten.py
