+++ game/src/gui/discard_pile.py
@@ -1,6 +1,6 @@
 import pygame
 
-import game.gui.tile_sprite as tile_sprite
+import game.src.gui.tile_sprite as tile_sprite
 
 
 class DiscardPile:
