+++ game/src/gui/resource_manager.py
