+++ main.py
@@ -1,4 +1,4 @@
-from game.gui.game import Game
+from game import Game
 
 
 if __name__ == "__main__":
