+++ ml/src/models/compare_models.py
@@ -1,7 +1,7 @@
 import random
 from operator import add
 
-from game.core.game_logic import simulate_match
+from game.src.core.game_logic import simulate_match
 
 
 def versus(competitors, how_many, init_seed, device):
